# lib/api-zod/src/generated/types/storyTemplateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mawaeedak API
 * OpenAPI spec version: 0.1.0
 */

export interface StoryTemplateUpdate {
  name?: string;
  description?: string;
  template_text?: string;
  background_color?: string;
  text_color?: string;
  is_active?: boolean;
}